backend_examples/requirements.txt
aiofiles==0.4.0
aiopg==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.2.5
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.4
pkg-resources==0.0.0
psycopg2-binary==2.8.4
rfc3986==1.3.2
sanic==19.12.2
sniffio==1.1.0
ujson==1.35
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.0
